# Optimize ggcr interactions