This is a first pass at speeding up the registry portion of apko publish.

```
hyperfine --warmup 1 --runs 3 "apko publish examples/nginx.yaml $APKO_REPO" "./apko publish examples/nginx.yaml $APKO_REPO"
Benchmark 1: apko publish examples/nginx.yaml $APKO_REPO
  Time (mean ± σ):     22.535 s ±  0.478 s    [User: 5.212 s, System: 3.160 s]
  Range (min … max):   22.097 s … 23.046 s    3 runs

Benchmark 2: ./apko publish examples/nginx.yaml $APKO_REPO
  Time (mean ± σ):      8.323 s ±  0.200 s    [User: 4.510 s, System: 2.379 s]
  Range (min … max):    8.193 s …  8.554 s    3 runs

Summary
  './apko publish examples/nginx.yaml $APKO_REPO' ran
    2.71 ± 0.09 times faster than 'apko publish examples/nginx.yaml $APKO_REPO'
```