@jonjohnsonjr 2023-04-25
> Do you think that most of the speed-up came from errgroups, or reusing pushers?

Mostly groups:

```
hyperfine --warmup 1 --runs 3 "apko publish examples/nginx.yaml $APKO_REPO" "./apko publish examples/nginx.yaml $APKO_REPO"
Benchmark 1: apko publish examples/nginx.yaml $APKO_REPO
  Time (mean ± σ):     22.448 s ±  1.025 s    [User: 5.148 s, System: 3.399 s]
  Range (min … max):   21.673 s … 23.611 s    3 runs

Benchmark 2: ./apko publish examples/nginx.yaml $APKO_REPO
  Time (mean ± σ):     16.222 s ±  0.627 s    [User: 4.459 s, System: 2.679 s]
  Range (min … max):   15.724 s … 16.926 s    3 runs

Summary
  './apko publish examples/nginx.yaml $APKO_REPO ran
    1.38 ± 0.08 times faster than 'apko publish examples/nginx.yaml $APKO_REPO
```

https://github.com/jonjohnsonjr/apko/commit/79d5a211ba1c473482996d57fbee7ebb9015e81e